requirements.txt
# Needs to install:

# django
Django==3.0.3
# for postgresql
gunicorn==20.0.4
psycopg2-binary==2.8.4
# for autodoc
Sphinx==2.4.0
yummy-sphinx-theme==0.1.1
# for api
djangorestframework==3.11.0
djoser==2.0.3
djangorestframework-jwt==1.11.0